requirements.txt
streamlit>=1.36
torch>=2.2
torchvision>=0.17
Pillow>=10.0
numpy>=1.26
matplotlib>=3.8
pandas>=2.2
scikit-learn>=1.4
